add more tests, add duo
