+++ authentik/stages/authenticator_sms/models.py
@@ -170,6 +170,12 @@ class SMSDevice(SideChannelDevice):
 
     last_t = models.DateTimeField(auto_now=True)
 
+    def verify_token(self, token):
+        valid = super().verify_token(token)
+        if valid:
+            self.save()
+        return valid
+
     def __str__(self):
         return self.name or str(self.user)
